ai_assitant/backend/requirements.txt

# backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.6.0

